backend/requirements.txt
fastapi==0.115.5
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.6.1
pydantic[email]==2.10.2
python-dotenv==1.0.1
uvicorn[standard]==0.32.1
pytest==8.3.4
httpx==0.28.0
